+++ cli/cmd/sync_v3.go
@@ -34,6 +34,7 @@ import (
 	"github.com/cloudquery/cloudquery/cli/v6/internal/tablenamechanger"
 	"github.com/cloudquery/cloudquery/cli/v6/internal/transformer"
 	"github.com/cloudquery/cloudquery/cli/v6/internal/transformerpipeline"
+	"github.com/cloudquery/cloudquery/cli/v6/internal/utils"
 )
 
 type v3source struct {
@@ -149,7 +150,7 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 		syncTimeTook   time.Duration
 		totalResources = int64(0)
 		totals         = sourceClient.Metrics()
-		statsPerTable  = cloudquery_api.SyncRunTableProgress{}
+		statsPerTable  = utils.NewConcurrentMap[string, cloudquery_api.SyncRunTableProgressValue]()
 	)
 	defer func() {
 		analytics.TrackSyncCompleted(ctx, invocationUUID.UUID, analytics.SyncFinishedEvent{
@@ -363,10 +364,11 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 	}
 	// Pre init stats per table
 	for _, table := range sourceTables {
-		statsPerTable[table.Name] = cloudquery_api.SyncRunTableProgressValue{
+		initialStats := cloudquery_api.SyncRunTableProgressValue{
 			Rows:   0,
 			Errors: 0,
 		}
+		statsPerTable.Add(table.Name, initialStats)
 	}
 
 	var remoteProgressReporter *godebouncer.Debouncer
@@ -379,7 +381,8 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 		if err != nil {
 			return fmt.Errorf("failed to parse sync run ID: %w", err)
 		}
-		remoteProgressReporter = godebouncer.NewWithOptions(godebouncer.WithTimeDuration(10*time.Second), godebouncer.WithTriggered(func() {
+
+		triggerFunc := func() {
 			totals := sourceClient.Metrics()
 			for i := range destinationsClients {
 				m := destinationsClients[i].Metrics()
@@ -390,12 +393,13 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 			if atomic.LoadInt64(&isComplete) == 1 {
 				status = cloudquery_api.SyncRunStatusCompleted
 			}
+			tableProgress := cloudquery_api.SyncRunTableProgress(statsPerTable.GetAll())
 			obj := cloudquery_api.CreateSyncRunProgressJSONRequestBody{
 				Rows:          atomic.LoadInt64(&totalResources),
 				Errors:        int64(totals.Errors),
 				Warnings:      int64(totals.Warnings),
 				Status:        &status,
-				TableProgress: &statsPerTable,
+				TableProgress: &tableProgress,
 			}
 			if shard != nil {
 				obj.ShardNum = &shard.num
@@ -411,8 +415,13 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 			if res.StatusCode() != http.StatusNoContent {
 				log.Warn().Str("status", res.Status()).Int("code", res.StatusCode()).Msg("Failed to send sync progress to API")
 			}
-		}), godebouncer.WithOptions(godebouncer.Options{Trailing: true, Leading: true}))
-		defer remoteProgressReporter.Cancel()
+		}
+		remoteProgressReporter = godebouncer.NewWithOptions(godebouncer.WithTimeDuration(10*time.Second), godebouncer.WithTriggered(triggerFunc), godebouncer.WithOptions(godebouncer.Options{Trailing: true, Leading: true}))
+
+		defer func() {
+			remoteProgressReporter.Cancel()
+			triggerFunc()
+		}()
 	}
 
 	// Transformers can change table names. We need to keep track of table name changes
@@ -507,9 +516,9 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 				atomic.AddInt64(&newResources, record.NumRows())
 				atomic.AddInt64(&totalResources, record.NumRows())
 				tableName, _ := record.Schema().Metadata().GetValue(schema.MetadataTableName)
-				stats := statsPerTable[tableName]
+				stats, _ := statsPerTable.Get(tableName)
 				stats.Rows += record.NumRows()
-				statsPerTable[tableName] = stats
+				statsPerTable.Add(tableName, stats)
 				if remoteProgressReporter != nil {
 					remoteProgressReporter.SendSignal()
 				}
@@ -623,9 +632,9 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 				}
 			case *plugin.Sync_Response_Error:
 				log.Error().Str("table", m.Error.TableName).Msg(m.Error.Error)
-				stats := statsPerTable[m.Error.TableName]
+				stats, _ := statsPerTable.Get(m.Error.TableName)
 				stats.Errors++
-				statsPerTable[m.Error.TableName] = stats
+				statsPerTable.Add(m.Error.TableName, stats)
 			default:
 				return fmt.Errorf("unknown message type: %T", m)
 			}
@@ -643,13 +652,15 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 	totals = sourceClient.Metrics()
 	sourceWarnings := totals.Warnings
 	var sourceErrors uint64
-	for _, val := range statsPerTable {
+	for _, val := range statsPerTable.GetAll() {
 		sourceErrors += uint64(val.Errors)
 	}
 	if totals.Errors > sourceErrors {
 		sourceErrors = totals.Errors
 	}
 	var metadataDataErrors error
+
+	tableProgress := statsPerTable.GetAll()
 	for i := range destinationsClients {
 		m := destinationsClients[i].Metrics()
 		summary := syncSummary{
@@ -668,6 +679,14 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 			DestinationName:     destinationSpecs[i].Name,
 			DestinationVersion:  destinationSpecs[i].Version,
 			DestinationPath:     destinationSpecs[i].Path,
+			ResourcesPerTable: lo.Reduce(lo.Keys(tableProgress), func(acc map[string]uint64, tableName string, _ int) map[string]uint64 {
+				acc[tableName] = uint64(tableProgress[tableName].Rows)
+				return acc
+			}, map[string]uint64{}),
+			ErrorsPerTable: lo.Reduce(lo.Keys(tableProgress), func(acc map[string]uint64, tableName string, _ int) map[string]uint64 {
+				acc[tableName] = uint64(tableProgress[tableName].Errors)
+				return acc
+			}, map[string]uint64{}),
 		}
 
 		if destinationSpecs[i].SyncGroupId != "" {
@@ -729,10 +748,6 @@ func syncConnectionV3(ctx context.Context, syncOptions syncV3Options) (syncErr e
 		Str("result", msg).
 		Msg("Sync summary")
 
-	if remoteProgressReporter != nil {
-		remoteProgressReporter.SendSignal()
-		<-remoteProgressReporter.Done()
-	}
 	return nil
 }