+++ plugins/destination/clickhouse/queries/tables_test.go
@@ -27,7 +27,7 @@ func TestCreateTable(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "", spec.DefaultEngine(), nil, nil)
+	}, "", spec.DefaultEngine(), nil, nil, nil)
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table.sql")
 }
@@ -48,10 +48,11 @@ func TestCreateTableWithPartitionBy(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "", spec.DefaultEngine(), []spec.PartitionStrategy{{Tables: []string{"*"}, PartitionBy: "toYYYYMM(`_cq_sync_time`)"}}, nil)
+	}, "", spec.DefaultEngine(), []spec.PartitionStrategy{{Tables: []string{"*"}, PartitionBy: "toYYYYMM(`_cq_sync_time`)"}}, nil, nil)
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table_partition_by.sql")
 }
+
 func TestCreateTableWithPartitionByErrorsIfTwoMatching(t *testing.T) {
 	query, err := CreateTable(&schema.Table{
 		Name: "table_name",
@@ -71,7 +72,7 @@ func TestCreateTableWithPartitionByErrorsIfTwoMatching(t *testing.T) {
 	}, "", spec.DefaultEngine(), []spec.PartitionStrategy{
 		{Tables: []string{"*"}, PartitionBy: "toYYYYMM(`_cq_sync_time`)"},
 		{Tables: []string{"*"}, PartitionBy: "toYYYYMM(`_cq_sync_time`)"},
-	}, nil)
+	}, nil, nil)
 	require.Error(t, err)
 	require.Empty(t, query)
 }
@@ -92,7 +93,7 @@ func TestCreateTableWithPartitionBySkipsIfNoMatch(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "", spec.DefaultEngine(), []spec.PartitionStrategy{{Tables: []string{"*"}, SkipTables: []string{"table_name"}, PartitionBy: "toYYYYMM(`_cq_sync_time`)"}}, nil)
+	}, "", spec.DefaultEngine(), []spec.PartitionStrategy{{Tables: []string{"*"}, SkipTables: []string{"table_name"}, PartitionBy: "toYYYYMM(`_cq_sync_time`)"}}, nil, nil)
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table.sql")
 }
@@ -114,7 +115,7 @@ func TestCreateTableWithPartitionBySkipsIfIncremental(t *testing.T) {
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
 		IsIncremental: true,
-	}, "", spec.DefaultEngine(), []spec.PartitionStrategy{{Tables: []string{"*"}, SkipIncrementalTables: true, PartitionBy: "toYYYYMM(`_cq_sync_time`)"}}, nil)
+	}, "", spec.DefaultEngine(), []spec.PartitionStrategy{{Tables: []string{"*"}, SkipIncrementalTables: true, PartitionBy: "toYYYYMM(`_cq_sync_time`)"}}, nil, nil)
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table.sql")
 }
@@ -135,10 +136,11 @@ func TestCreateTableWithOrderBy(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "", spec.DefaultEngine(), nil, []spec.OrderByStrategy{{Tables: []string{"table_name"}, OrderBy: []string{"`_cq_sync_time`", "`_cq_id`"}}})
+	}, "", spec.DefaultEngine(), nil, []spec.OrderByStrategy{{Tables: []string{"table_name"}, OrderBy: []string{"`_cq_sync_time`", "`_cq_id`"}}}, nil)
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table_order_by.sql")
 }
+
 func TestCreateTableWithOrderByErrorsIfTwoMatching(t *testing.T) {
 	query, err := CreateTable(&schema.Table{
 		Name: "table_name",
@@ -158,7 +160,7 @@ func TestCreateTableWithOrderByErrorsIfTwoMatching(t *testing.T) {
 	}, "", spec.DefaultEngine(), nil, []spec.OrderByStrategy{
 		{Tables: []string{"table_name"}, OrderBy: []string{"`_cq_sync_time`", "`_cq_id`"}},
 		{Tables: []string{"table_name"}, OrderBy: []string{"`_cq_sync_time`", "`_cq_id`"}},
-	})
+	}, nil)
 	require.Error(t, err)
 	require.Empty(t, query)
 }
@@ -179,7 +181,7 @@ func TestCreateTableWithOrderBySkipsIfNoMatch(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "", spec.DefaultEngine(), nil, []spec.OrderByStrategy{{Tables: []string{"table_name"}, SkipTables: []string{"table_name"}, OrderBy: []string{"`_cq_sync_time`", "`_cq_id`"}}})
+	}, "", spec.DefaultEngine(), nil, []spec.OrderByStrategy{{Tables: []string{"table_name"}, SkipTables: []string{"table_name"}, OrderBy: []string{"`_cq_sync_time`", "`_cq_id`"}}}, nil)
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table.sql")
 }
@@ -192,7 +194,7 @@ func TestCreateTableNoOrderBy(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "", spec.DefaultEngine(), nil, nil)
+	}, "", spec.DefaultEngine(), nil, nil, nil)
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table_empty_order_by.sql")
 }
@@ -213,7 +215,7 @@ func TestCreateTableOnCluster(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "my_cluster", spec.DefaultEngine(), nil, nil)
+	}, "my_cluster", spec.DefaultEngine(), nil, nil, nil)
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table_cluster.sql")
 }
@@ -237,9 +239,67 @@ func TestCreateTableWithEngine(t *testing.T) {
 	}, "", &spec.Engine{
 		Name:       "ReplicatedMergeTree",
 		Parameters: []any{"a", "b", 1, int32(2), int64(3), float32(1.2), float64(3.4), json.Number("327"), false, true},
-	}, nil, nil)
+	}, nil, nil, nil)
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table_engine.sql")
+}
+
+func TestCreateTableWithTTL_NotNullCqSyncTime(t *testing.T) {
+	notNullCqSyncTime := schema.CqSyncTimeColumn
+	notNullCqSyncTime.NotNull = true
+	ttlStrategy := []spec.TTLStrategy{
+		{
+			Tables:     []string{"table_*"},
+			SkipTables: nil,
+			TTL:        "INTERVAL 1 DAY + INTERVAL 5415 SECOND",
+		},
+	}
+	query, err := CreateTable(&schema.Table{
+		Name: "table_name",
+		Columns: schema.ColumnList{
+			schema.CqIDColumn,
+			schema.CqParentIDColumn,
+			schema.CqSourceNameColumn,
+			notNullCqSyncTime,
+			schema.Column{
+				Name:    "extra_col",
+				Type:    arrow.PrimitiveTypes.Float64,
+				NotNull: true,
+			},
+			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
+			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
+		},
+	}, "", spec.DefaultEngine(), nil, nil, ttlStrategy)
+	require.NoError(t, err)
+	ensureContents(t, query, "create_table_ttl_not_null_cq_time.sql")
+}
+
+func TestCreateTableWithTTL_NullableCqSyncTime(t *testing.T) {
+	ttlStrategy := []spec.TTLStrategy{
+		{
+			Tables:     []string{"table_*"},
+			SkipTables: nil,
+			TTL:        "INTERVAL 1 DAY + INTERVAL 5415 SECOND",
+		},
+	}
+	query, err := CreateTable(&schema.Table{
+		Name: "table_name",
+		Columns: schema.ColumnList{
+			schema.CqIDColumn,
+			schema.CqParentIDColumn,
+			schema.CqSourceNameColumn,
+			schema.CqSyncTimeColumn,
+			schema.Column{
+				Name:    "extra_col",
+				Type:    arrow.PrimitiveTypes.Float64,
+				NotNull: true,
+			},
+			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
+			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
+		},
+	}, "", spec.DefaultEngine(), nil, nil, ttlStrategy)
+	require.NoError(t, err)
+	ensureContents(t, query, "create_table_ttl.sql")
 }
 
 func TestDropTable(t *testing.T) {